python sqlAlchemy

+++ 0x0F-python-object_relational_mapping/13-model_state_delete_a.py
@@ -1,48 +1,33 @@
 #!/usr/bin/python3
-import sys
+"""
+This script will delete all State objects
+with a name containing the letter `a`
+from the database `hbtn_0e_6_usa`.
+"""
+
+from sys import argv
+from model_state import State, Base
 from sqlalchemy import create_engine
 from sqlalchemy.orm import sessionmaker
-from model_state import Base, State
-
-def delete_states_with_letter_a(username, password, hbtn_0e_6_usa):
-    # Create engine to connect to MySQL server
-    engine = create_engine(f"mysql+mysqldb://{username}:{password}@localhost:3306/{hbtn_0e_6_usa}")
-
-    # Create session factory
-    Session = sessionmaker(bind=engine)
-
-    # Create a session
-    session = Session()
-
-    try:
-        # Query to get all State objects with names containing 'a'
-        states_with_a = session.query(State).filter(State.name.ilike('%a%')).all()
-
-        # Delete the states
-        for state in states_with_a:
-            session.delete(state)
-
-        # Commit the session to save the changes to the database
-        session.commit()
-
-        print("Deleted states with 'a' in the name.")
-
-    except Exception as e:
-        # Rollback the session in case of an error
-        session.rollback()
-        print("Error:", e)
-
-    finally:
-        # Close the session
-        session.close()
 
 if __name__ == "__main__":
-    if len(sys.argv) != 4:
-        print("Usage: python script.py <username> <password> <database_name>")
-        sys.exit(1)
+    """
+    Deletes State objects on the database.
+    """
 
-    username = sys.argv[1]
-    password = sys.argv[2]
-    database_name = sys.argv[3]
+    db_url = "mysql+mysqldb://{}:{}@localhost:3306/{}".format(
+        argv[1], argv[2], argv[3])
 
-    delete_states_with_letter_a(username, password, database_name)
+    engine = create_engine(db_url)
+    Session = sessionmaker(bind=engine)
+
+    session = Session()
+
+    states = session.query(State).filter(State.name.contains('a'))
+    if states is not None:
+        for state in states:
+            session.delete(state)
+
+    session.commit()
+
+    session.close()
